### 2025-05-09 #8185 @editorialbot
[PRE REVIEW]: pyROX: Rapid Opacity X-sections
**Submitting author:** @samderegt (<a href="http://orcid.org/0000-0003-4760-6168">Sam de Regt</a>)
**Repository:** https://github.com/samderegt/pyROX
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
