### 2025-05-09 #8185 @warrickball
Hi @samderegt, thanks for your submission! I think the basic requirements already look in order with this submission. It's a minor point for later but your bibliography uses the NASA ADS shortcuts for journal names, which JOSS doesn't recognise. If you could replace e.g. `\apj` with "The Astrophysical Journal", the journal names will appear correctly in the references.

All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. The AASS track currently has a significant backlog so I'll warn you now that it might be several months before we're able to start processing this. Thanks for your patience!
