### 2025-05-12 9c20452
Replaced journal shortcuts with full names
