### 2025-05-12 #8185 @samderegt
Hi @warrickball, thanks for letting me know about the expected timeframe. I have updated the journal names in the bibliography, thank you for pointing this out.
