### 2025-09-13 db042b5
Add DOI badge to README
Added DOI badge to README for citation purposes.
