### 2025-10-22 5a3c450
Fixed installation of supplemental data files
