+++ setup.py
@@ -19,6 +19,10 @@ setup(
     long_description_content_type="text/markdown",
     url="https://github.com/samderegt/pyROX",  # Update with your repository URL
     packages=find_packages(),  # Automatically find package directories
+    include_package_data=True, 
+    package_data={
+        "pyROX": ["data/*","data/*/*"],  # Include data files in the package
+    },
     install_requires=read_requirements(),  # Read dependencies from requirements.txt
     classifiers=[
         "Programming Language :: Python :: 3",
