### 2026-02-03 fbd8473
New JOSS paper requirements
