+++ docs/JOSS/paper.md
@@ -35,7 +35,7 @@ affiliations:
     index: 2
   - name: Centre for Exoplanets and Habitability, University of Warwick, Gibbet Hill Road, Coventry CV4 7AL, UK
     index: 3
-date: 30 April 2025
+date: 02 February 2026
 bibliography: paper.bib
 
 ---
@@ -51,27 +51,35 @@ The advent of a new generation of telescopes and instruments has led to a dramat
 
 Opacity cross-sections play a key role in accurately modelling sub-stellar atmospheres. Opacity governs the dominant energy transport mechanism (i.e. radiation or convection) which affects the thermal structure of the atmosphere [@Marley_ea_2021]. Furthermore, high-resolution studies require well-determined frequencies for the transition lines. Inaccuracies in line-list data can result in biased abundance constraints [e.g. @Brogi_ea_2019; @de_Regt_ea_2024] or ambiguous (non)-detections of certain molecules [e.g. @de_Regt_ea_2022; @Merritt_ea_2020; @Serindag_ea_2021]. It is therefore important that the most up-to-date and complete opacity data are used. However, it can be difficult to efficiently calculate opacity cross-sections from line lists that sometimes consist of billions of transitions. 
 
-To help resolve this challenge, we present `pyROX`, a user-friendly Python package to calculate molecular and atomic cross-sections for applications in models of sub-stellar atmospheres. `pyROX` supports line opacity calculations from the ExoMol [@Tennyson_ea_2024], HITRAN [@Gordon_ea_2022], HITEMP [@Rothman_ea_2010], and Kurucz[^1] databases. Collision-Induced Absorption (CIA) coefficients can also be calculated from the HITRAN and Borysow[^2] databases. So far, `pyROX`-computed cross-sections have enabled several recent publications from our research group [@de_Regt_ea_2025; @Siebenaler_ea_2025].
+To help resolve this challenge, we present `pyROX`, a user-friendly Python package to calculate molecular and atomic cross-sections for applications in models of sub-stellar atmospheres. `pyROX` supports line opacity calculations from the ExoMol [@Tennyson_ea_2024], HITRAN [@Gordon_ea_2022], HITEMP [@Rothman_ea_2010], and Kurucz[^1] databases. Collision-Induced Absorption (CIA) coefficients can also be calculated from the HITRAN and Borysow[^2] databases. 
 
 [^1]: [http://kurucz.harvard.edu/](http://kurucz.harvard.edu/)
 [^2]: [https://www.astro.ku.dk/~aborysow/programs/index.html](https://www.astro.ku.dk/~aborysow/programs/index.html)
 
 
-# Functionality of `pyROX`
-Documentation for `pyROX` is available at [https://py-rox.readthedocs.io/en/latest/](https://py-rox.readthedocs.io/en/latest/) and includes tutorial examples to running the code. Here, we outline the main functionality of `pyROX`:
-
-- **Download and read files**: The necessary input files (line lists, partition functions, broadening coefficients or CIA files) can be downloaded with a simple command. When reading the relevant parameters, `pyROX` handles the different data structures of the supported databases (ExoMol, HITRAN/HITEMP, Kurucz, Borysow). 
-- **Compute line-strengths and -widths**: For line-opacity calculations, `pyROX` calculates the strength and broadening-widths for each line transition at the user-provided pressure and temperature. Support is offered for various [pressure-broadening descriptions](https://py-rox.readthedocs.io/en/latest/notebooks/pressure_broadening.html). `pyROX` can speed up the line-profile computation by selecting only the main line-strength contributors.
-- **Compute line profiles**: Next, `pyROX` computes the Voigt profiles as the real part of the Faddeeva function [Eq. 12 of @Gandhi_ea_2020], using the `scipy.special.wofz` implementation[^3]. 
[…]
