+++ docs/JOSS/paper.bib
@@ -197,6 +197,42 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Gonzalez_Picos_ea_2025b,
+       author = {{Gonz{\'a}lez Picos}, D. and {de Regt}, S. and {Gandhi}, S. and {Grasser}, N. and {Snellen}, I.~A.~G.},
+        title = "{Disentangling disc and atmospheric signatures of young brown dwarfs with JWST/NIRSpec}",
+      journal = {\aap},
+     keywords = {techniques: spectroscopic, planets and satellites: atmospheres, brown dwarfs, Earth and Planetary Astrophysics, Solar and Stellar Astrophysics},
+         year = 2025,
+        month = nov,
+       volume = {703},
+          eid = {A65},
+        pages = {A65},
+          doi = {10.1051/0004-6361/202555669},
+archivePrefix = {arXiv},
+       eprint = {2509.11694},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025A&A...703A..65G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Gonzalez_Picos_ea_2025c,
+       author = {{Gonz{\'a}lez Picos}, Dar{\'\i}o and {Snellen}, Ignas and {de Regt}, Sam},
+        title = "{Chemical evolution imprints in the rare isotopes of nearby M dwarfs}",
+      journal = {Nature Astronomy},
+     keywords = {Physical Sciences, Astronomical and Space Sciences, Solar and Stellar Astrophysics, Earth and Planetary Astrophysics, Astrophysics of Galaxies},
+         year = 2025,
+        month = nov,
+       volume = {9},
+       number = {11},
+        pages = {1692-1700},
+          doi = {10.1038/s41550-025-02641-4},
+archivePrefix = {arXiv},
+       eprint = {2508.18424},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025NatAs...9.1692G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Gordon_ea_2022,
[…]
