### 2026-02-03 d61d60d
Replaced journal shortcuts with full names
