+++ docs/JOSS/paper.bib
@@ -201,7 +201,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Gonzalez_Picos_ea_2025b,
        author = {{Gonz{\'a}lez Picos}, D. and {de Regt}, S. and {Gandhi}, S. and {Grasser}, N. and {Snellen}, I.~A.~G.},
         title = "{Disentangling disc and atmospheric signatures of young brown dwarfs with JWST/NIRSpec}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {techniques: spectroscopic, planets and satellites: atmospheres, brown dwarfs, Earth and Planetary Astrophysics, Solar and Stellar Astrophysics},
          year = 2025,
         month = nov,
@@ -451,7 +451,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Schreier_2018,
        author = {{Schreier}, Franz},
         title = "{The Voigt and complex error function: Huml{\'\i}{\v{c}}ek's rational approximation generalized}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {line: profiles, methods: numerical, techniques: spectroscopic, Physics - Computational Physics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = sep,
@@ -505,7 +505,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Siebenaler_ea_2025b,
        author = {{Siebenaler}, Louis and {Miguel}, Yamila},
         title = "{Mean Opacity Tables for Probing the Interior and Atmosphere of Giant Planets}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {Earth and Planetary Astrophysics, Solar and Stellar Astrophysics},
          year = 2025,
         month = dec,
@@ -556,7 +556,7 @@ archivePrefix = {arXiv},
 @ARTICLE{van_Sluijs_ea_2025,
        author = {{van Sluijs}, Lennart and {Rauscher}, Emily and {Kempton}, Eliza M.-R. and {Kennedy}, Thomas and {Malsky}, Isaac and {Matsunaga}, Noriyuki and {Meyer}, Michael and {McWilliam}, Andrew and {Monnier}, John D. and {Otsubo}, Shogo and {Sarugaku}, Yuki and {Takeuchi}, Tomomi},
         title = "{First Results from WINERED: Detection of Emission Lines from Neutral Iron and a Combined Set of Trace Species on the Dayside of WASP-189 b}",
-      journal = {\aj},
+      journal = {Astronomical Journal},
      keywords = {Exoplanet atmospheric structure, High resolution spectroscopy, Exoplanet atmospheric composition, 2310, 2096, 2021, Earth and Planetary Astrophysics, Instrumentation and Methods for Astrophysics},
          year = 2025,
         month = oct,
@@ -666,7 +666,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Zhang_ea_2026,
        author = {{Zhang}, Yapeng and {Wardenier}, Joost P. and {Householder}, Aaron and {Komacek}, Thaddeus D. and {Kesseli}, Aurora Y. and {Dai}, Fei and {Howard}, Andrew W. and {Inglis}, Julie and {Isaacson}, Howard and {Knutson}, Heather A. and {Mawet}, Dimitri and {Pino}, Lorenzo and {Wallack}, Nicole and {Xuan}, Jerry W. and {Carmichael}, Theron W. and {Huber}, Daniel and {Lee}, Rena A. and {Saunders}, Nicholas and {Weiss}, Lauren and {Zhang}, Jingwen},
[…]
