### 2026-02-04 #8185 @samaloney
Hi @samderegt, sorry for the very long wait and thanks for getting the submission updated for JOSS’s new scope requirements. I was just wondering if you had any suggestions for potential reviewers, as finding potential reviews can often be the next most time-consuming step? If you do then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
