### 2026-02-06 #8185 @samderegt
Hi @samaloney, thank you for picking this up. Looking through the list of potential reviewers, I jotted down the names of Arjun Savel, Ben Lew, Evert Nasedkin, and Ehsan (Sam) Gharib-Nezhad -- if they are still up for reviewing JOSS submissions of course.
