### 2026-03-09 #10081 @samaloney
Hey @arjunsavel, @kyleaoman, @ttricco I'm checking in to see how the review is going?

Don't forget you can generate your checklist with `@editorialbot generate my checklist`. The review doesn't have to be completed in one go, you can work through each item as time permits if that's better for you.
