### 2026-03-09 #10081 @arjunsavel
@samaloney thanks for the nudge! I'm making my way through the paper now.

For the COI, just want to flag that I've been on a [few publications](https://ui.adsabs.harvard.edu/search/filter_property_fq_property=AND&filter_property_fq_property=property%3A%22refereed%22&fq=%7B!type%3Daqp%20v%3D%24fq_database%7D&fq=%7B!type%3Daqp%20v%3D%24fq_property%7D&fq_database=(database%3Aastronomy%20OR%20database%3Aphysics)&fq_property=(property%3A%22refereed%22)&q=%20author%3A%22Savel%2C%20Arjun%22%2C%20%22Gandhi%2C%20Siddharth%22&sort=date%20desc%2C%20bibcode%20desc&p_=0) with the second author. The first is a paper the second author led in 2022 (I was last author), and the second is a paper in 2025 in which he was 5th author and I was 9th. My personal assessment is that I can provide an impartial review, but I'm happy to discuss this.
