### 2026-03-17 samderegt/pyROX#20 @kyleaoman
Supported python & dependency versions are not specific enough
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/10081).

I encounter an error that comes from an incompatible `numpy` version. I am running:
```
Python 3.14.3 (main, Feb  4 2026, 00:00:00) [GCC 15.2.1 20260123 (Red Hat 15.2.1-7)] on linux
```
I set up a clean virtual environment then installed `pyROX`, and whatever dependencies it automatically installed (plus `ipykernel` to make the kernel available in notebooks). The exact package listing is:
```
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
h5py==3.16.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
pyROX @ file:///home/txwx36/code/pyROX
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.32.5
scipy==1.17.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
urllib3==2.6.3
wcwidth==0.6.0
wget==3.2
```
With this, when I reach the following cell in `getting_started.ipynb`:
```
data.calculate_temporary_outputs(
    progress_bar=True, # Show progress bar
    #overwrite=False, # If True, answer "yes" to all overwrite prompts
)
```
an exception is raised:
```
Cell In[10], line 1
----> 1 data.calculate_temporary_outputs(
      2     progress_bar=True, # Show progress bar
      3     #overwrite=False, # If True, answer "yes" to all overwrite prompts
      4 )

File ~/.venv/pyrox/lib64/python3.14/site-packages/pyROX/line_by_line/lbl.py:872, in LineByLine.calculate_temporary_outputs(self, overwrite, save_in_one_file, files_range, **kwargs)
    866 self.sigma = np.zeros(
    867     (len(self.nu_grid), len(self.P_grid), len(self.T_grid)), dtype=float
    868     )
    869 for input_file, tmp_output_file in zip(transitions_files, tmp_output_files):
    870
    871     # Compute the cross-sections
--> 872     self.process_transitions(input_file, **kwargs)
    874     if np.all(self.sigma == 0.):
    875         print(f'  No lines calculated, no need to save \"{tmp_output_file}\"')

File ~/.venv/pyrox/lib64/python3.14/site-packages/pyROX/line_by_line/lbl_exomol.py:294, in LBL_ExoMol.process_transitions(self, input_file, **kwargs)
    292 # Compute the cross-sections, looping over the PT-grid
    293 print(f'  Number of lines loaded: {len(A)}')
--> 294 self.iterate_over_PT_grid(
    295     function=self.calculate_cross_sections,
    296     nu_0=nu_0, S_0=S_0, E_low=E_low, A=A, **kwargs
    297 )
    299 # Reset
    300 state_ID_up = []; state_ID_low = []; A = []

File ~/.venv/pyrox/lib64/python3.14/site-packages/pyROX/line_by_line/lbl.py:750, in LineByLine.iterate_over_PT_grid(self, function, progress_bar, **kwargs)
    747 for idx_P, P in enumerate(self.P_grid):
    748     for idx_T, T in enumerate(self.T_grid):
--> 750         function(P, T, **kwargs)
    751         postfix = {
    752             'P': f'{P*1e-5:.0e} bar', 'T': f'{T:.0f} K',
    753         }
    754         if hasattr(self, 'N_lines_computed'):

File ~/.venv/pyrox/lib64/python3.14/site-packages/pyROX/line_by_line/lbl.py:822, in LineByLine.calculate_cross_sections(self, P, T, nu_0, S_0, E_low, A, delta, **kwargs)
    819 S = self.normalise_wing_cutoff(S, wing_cutoff_distance, gamma_L)
    821 # Compute the line-profiles in chunks
--> 822 sigma = self.calculate_line_profiles(
    823     nu_0=nu_0,
    824     S=S,
    825     gamma_L=gamma_L,
    826     gamma_G=gamma_G,
    827     nu_grid=nu_grid_to_use,
    828     wing_cutoff_distance=wing_cutoff_distance,
    829     delta_nu=delta_nu_to_use,
    830 )
    832 if len(nu_grid_to_use) != len(self.nu_grid):
    833     # Interpolate to the original grid
    834     sigma = np.interp(self.nu_grid, nu_grid_to_use, sigma)

File ~/.venv/pyrox/lib64/python3.14/site-packages/pyROX/line_by_line/lbl.py:266, in LineProfileHelper.calculate_line_profiles(self, nu_0, S, gamma_L, gamma_G, nu_grid, wing_cutoff_distance, delta_nu)
    249 """
    250 Calculates line profiles in chunks to optimise speed.
    251
   (...)    262     array: Opacity cross-section on the grid.
    263 """
    264 if not np.isnan(self.delta_nu):
    265     # Fixed wavenumber spacing
--> 266     return self._calculate_line_profiles_in_chunks_fixed_delta_nu(
    267         nu_0, S, gamma_L, gamma_G, nu_grid, wing_cutoff_distance, delta_nu
    268         )
    270 # Other wavenumber grids
    271 return self._calculate_line_profiles_in_chunks(
    272     nu_0, S, gamma_L, gamma_G, nu_grid, wing_cutoff_distance, delta_nu
    273     )

File ~/.venv/pyrox/lib64/python3.14/site-packages/pyROX/line_by_line/lbl.py:433, in LineProfileHelper._calculate_line_profiles_in_chunks_fixed_delta_nu(self, nu_0, S, gamma_L, gamma_G, nu_grid, wing_cutoff_distance, delta_nu, N_lines_in_chunk)
    430 N_nu_grid = len(nu_grid)
    432 # Construct the temporary line-profile grid
--> 433 nu_line = np.arange(0, wing_cutoff_distance+delta_nu, delta_nu)
    434 nu_line = np.concatenate([-nu_line[1:][::-1], nu_line])
    435 nu_line = np.repeat(nu_line[None,:], N_lines_in_chunk, axis=0)

TypeError: only 0-dimensional arrays can be converted to Python scalars
```
I think that this comes from changes introduced in `numpy2.4`. I had a look at the CI configuration in `.github/workflows/python-tests.yml` and tests are done on `python3.8` and `python3.12`. The `setup.py` file notes `python_requires=">=3.7"`.

It's probably complicated to support all the way back to `python3.7` while also supporting current versions. Everything up to and including `python3.9` has reached EoL (https://devguide.python.org/versions/). I would thus suggest currently aiming to support `python3.10` through `python3.14` if possible. If this proves to be too complicated then picking a current version (could be `python3.12` that is currently tested) as a supported version makes sense. In any case more care should be taken to specify the supported versions of python and dependencies in the `setup.py` and `requirements.py` files. It is probably enough to just limit the supported python versions to the versions that are tested in CI.
