### 2026-03-17 d70f7a4
Simplified parallelisation through joblib
