+++ pyROX/cross_sections.py
@@ -417,6 +417,11 @@ class CrossSections:
         self.database = getattr(self.config, 'database', None)
         self.species  = getattr(self.config, 'species', None)
 
+        self.N_CPUs = getattr(self.config, 'N_CPUs', None)
+        if self.N_CPUs is None:
+            import os
+            self.N_CPUs = os.cpu_count()
+
         # Wavenumber/wavelength grid
         self._setup_nu_grid()
         
@@ -463,7 +468,7 @@ class CrossSections:
 
         self.delta_nu   = getattr(self.config, 'delta_nu', np.nan) # [cm^-1]
         self.delta_nu   = self.delta_nu * 1e2*sc.c # [s^-1]
-        self.delta_wave = getattr(self.config, 'self.delta_wave', np.nan) # [um]
+        self.delta_wave = getattr(self.config, 'delta_wave', np.nan) # [um]
         self.delta_wave = self.delta_wave * sc.micron # [um] -> [m]
         
         self.wave_min = getattr(self.config, 'wave_min', 1.0/3.0) # [um]
