### 2026-03-17 samderegt/pRT_opacities#21 @kyleaoman
Missing HITRAN files
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/10081).

When attempting to run the `collision_induced_absorption.ipynb` notebook, some files are not found (trying with my browser, I also find that the files are not accessible):
```
Downloading CIA data
  Downloading "https://hitran.org/data/CIA/H2-H2_2011.cia"
/cosma/home/durham/dc-oman1/.virtualenv/pyrox/lib/python3.12/site-packages/pyROX/utils.py:56 UserWarning: Failed to download "https://hitran.org/data/CIA/H2-H2_2011.cia": HTTP Error 404: Not Found
/cosma/home/durham/dc-oman1/.virtualenv/pyrox/lib/python3.12/site-packages/pyROX/utils.py:71 UserWarning: Failed to download "https://hitran.org/data/CIA/H2-H2_2011.cia": 404 - Not Found
  Downloading "https://hitran.org/data/CIA/supplementary/H2-H2_eq_2018.cia"
/cosma/home/durham/dc-oman1/.virtualenv/pyrox/lib/python3.12/site-packages/pyROX/utils.py:56 UserWarning: Failed to download "https://hitran.org/data/CIA/supplementary/H2-H2_eq_2018.cia": HTTP Error 404: Not Found
/cosma/home/durham/dc-oman1/.virtualenv/pyrox/lib/python3.12/site-packages/pyROX/utils.py:71 UserWarning: Failed to download "https://hitran.org/data/CIA/supplementary/H2-H2_eq_2018.cia": 404 - Not Found
```
Perhaps these have moved? Also possible that this is just a transient service outage.

I suggest running the notebooks in the CI pipeline. One possible mechanism to achieve this is illustrated here: https://github.com/SWIFTSIM/swiftgalaxy/blob/7c7ce2183d77c31a207b5d1448912821103370d5/tests/test_examples.py#L187 (this is the solution that I currently use for packages that I maintain, using the `nbmake` package, but I would note that I've found `nbmake` a bit awkward to use, so exploring the available options and choosing the one that you find most suitable is a good idea).
