### 2026-03-17 samderegt/pRT_opacities#22 @kyleaoman
Import errors in pressure broadening notebook & getting started example (command line)
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/10081).

This import: `import examples.exomol_alh.exomol_alh as config` fails with `ModuleNotFoundError: No module named 'examples.exomol_alh.exomol_alh'`. The working directory is `.../pyROX/docs/notebooks`, so it seems like it makes sense that this is not found (`examples` is in `.../pyROX/`). I can hack a work around with:
```
import sys
sys.path.append("../../")
```
and the rest of the notebook then runs, but a bit of re-thinking of the package structure & imports might be needed here.

I also get an error that seems related when trying to run the "getting started" example from the command line following the instructions in `getting_started.ipynb`:
```
dc-oman1@login8b> pwd
/cosma/home/durham/dc-oman1/code/pyROX
dc-oman1@login8b> ls
build  docs  examples  LICENSE  pyROX  pyROX.egg-info  README.md  requirements.txt  setup.py  tests
dc-oman1@login8b> pyROX examples/exomol_alh/exomol_alh.py

================================================================================
  Welcome to pyROX: Rapid Opacity X-sections for Python
================================================================================

Traceback (most recent call last):
  File "/cosma/home/durham/dc-oman1/.virtualenv/pyrox/bin/pyROX", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/cosma/home/durham/dc-oman1/.virtualenv/pyrox/lib/python3.12/site-packages/pyROX/main.py", line 71, in main
    config = __import__(config_string, fromlist=[''])
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
ModuleNotFoundError: No module named 'examples'
```
This is with: `Python 3.12.4 (main, Jul  4 2024, 14:42:26) [GCC 11.4.1 20231218 (Red Hat 11.4.1-3)] on linux` with a fresh pyROX installation on a clean virtual environment.
