### 2026-03-17 15e2820
Fixed bug encountered with numpy2.4 and typo for "delta_wave". Updated support to python>=3.10. Added weekly tests.
