### 2026-03-17 samderegt/pRT_opacities#23 @kyleaoman
Consider filtering warnings in tests
This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/10081).

When I run the test suite I get a large number (77) of warnings. `pyROX` uses warnings to pass messages to the user (perhaps a strange choice?). I'd suggest filtering warnings in the tests and ignoring those that are expected. It might be useful to create a custom `pyROXWarning` class for this - also useful for users to more easily catch/ignore these warnings if desired.
