### 2026-03-17 #10081 @kyleaoman
@samaloney I've completed an initial review and opened a few issues (linked above via mentions). Overall it's a well-presented package. Two general comments w.r.t. the two items I've left un-ticked on my checklist:
- There doesn't seem to be any evidence of community engagement (e.g. 0 issues before mine, 0 PRs). There are a couple of papers using & citing the package mentioned in the manuscript, but that seems to be about it. Leave it to you to evaluate what to do with that information.
- The test suite is quite limited, there are 4 tests that check more or less "end-to-end" results for correctness. Some of the issues that I opened could be seen as symptoms of insufficient test coverage, but I'm not sure how much I should push for a more comprehensive test suite as part of review. If that's something that's expected then I can open another issue asking for that.

Happy to have another look if/when needed!
