### 2026-03-18 7a37e32
Fixed bug encountered with numpy2.4 and typo for "delta_wave". Updated support to python>=3.10. Added weekly tests. (#24)
