+++ .github/workflows/python-tests.yml
@@ -10,6 +10,8 @@ on:
       - main
     paths-ignore:
       - 'docs/**'
+  schedule:
+    - cron: '0 8 * * 1' # Every Monday at 08:00 UTC
 
 jobs:
   test:
@@ -18,7 +20,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: ['3.8', '3.12']
+        python-version: ['3.10', '3.12', '3.14']
 
     steps:
     - name: Checkout code
