### 2026-03-18 1a17d3d
Custom pyROXWarning, to be ignored during tests
