+++ .github/workflows/python-tests.yml
@@ -39,4 +39,4 @@ jobs:
 
     - name: Run tests
       run: |
-        pytest --disable-warnings
+        pytest -W ignore::pyROX.utils.pyROXWarning
