+++ .github/workflows/python-tests.yml
@@ -35,8 +35,9 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -e .
-        pip install pytest
+        pip install pytest nbmake
 
     - name: Run tests
       run: |
         pytest -W ignore::pyROX.utils.pyROXWarning
+        pytest -W ignore::pyROX.utils.pyROXWarning --nbmake docs/notebooks/*.ipynb
