### 2026-03-18 2630cbd
More complete end-to-end integration tests
