+++ tests/test_lbl_cia.py
@@ -1,252 +1,0 @@
-import numpy as np
-import pathlib
-
-from pyROX import utils, cross_sections
-import shutil
-import sys
-
-# Add the root directory to the Python path
-parent_dir = pathlib.Path(__file__).resolve().parent
-sys.path.insert(0, str(parent_dir.parent))
-
-reference_data_dir = parent_dir / 'reference_data'
-reference_data_dir.mkdir(parents=True, exist_ok=True)
-
-input_data_dir  = parent_dir / 'input_data'
-output_data_dir = parent_dir / 'output_data'
-
-def test_lbl_kurucz():
-    """
-    Test the Kurucz line profile calculation.
-    """
-    # Filenames of synthetic data
-    input_data_dir.mkdir(parents=True, exist_ok=True)
-    states_file      = input_data_dir / 'kurucz_states.txt'
-    transitions_file = input_data_dir / 'kurucz_transitions.txt'
-
-    # Create the synthetic states file
-    states_data = [
-        '2\t\"\"\t\"0.0000\"\t\"\"',
-        '2\t\"\"\t\"12985.185724\"\t\"\"',
-        '4\t\"\"\t\"13042.896027\"\t\"\"',
-        '2\t\"\"\t\"21026.551\"\t\"\"',
-        '6\t\"\"\t\"21534.680\"\t\"\"',
-        '4\t\"\"\t\"21536.988\"\t\"\"',
-        '\t\"\"\t\"35009.8140\"\t\"\"'
-        
-    ]
-    with open(states_file, 'w') as f:
-        f.write("g\tPrefix\tLevel (cm-1)\tSuffix\n")
-        f.write("\n".join(states_data) + "\n")
-
-    # Create the synthetic transitions file
-    transition = '   100.0000 -2.503 19.00      50.000  0.5 4s 2S          100.000  1.5 3d3D4s 2D   9.41 -5.87 -7.72K12  0 0  0 0.000  0 0.000                     2002  801     0'
-    with open(transitions_file, 'w') as f:
-        f.write(transition + '\n')
-
-    # Import the example configuration
-    import examples.kurucz_k.kurucz_k as config
-    files = dict(states=states_file, transitions=transitions_file)
-    config = setup_lbl_config(config, files)
-
-    # Compare the line profiles to a reference file
-    compare_outputs(config, reference_data_dir/'kurucz_xsec.txt')
-   
-def test_lbl_hitran():
-    """
-    Test the HITRAN line profile calculation.
-    """
-    # Filenames of synthetic data
-    input_data_dir.mkdir(parents=True, exist_ok=True)
-    transitions_file = input_data_dir / 'hitran_transitions.txt'
-    partition_function_file = input_data_dir / 'hitran_pf.txt'
-
-    # Create the synthetic transition file
-    transition = ' 51  100.0000001.691E-031 1.284E-09.05550.061   50.00000.72-.024493             11              0                    R 13      447664 5 8 2 2 1 7    29.0   27.0'
-    with open(transitions_file, 'w') as f:
-        f.write(transition + '\n')
-
-    # Create the synthetic partition function
-    T = np.arange(0, 3000., 1)
-    with open(partition_function_file, 'w') as f:
-        for t in T:
-            f.write(f'{t} {2*t}\n')
-    
-    # Import the example configuration
-    import examples.hitemp_co.hitemp_co as config
-    files = dict(
-        transitions=transitions_file, 
-        partition_function=partition_function_file
-    )
-    config = setup_lbl_config(config, files)
-
-    # Compare the line profiles to a reference file
-    compare_outputs(config, reference_data_dir/'hitran_xsec.txt')
-
-def test_lbl_exomol():
-    """
[…]
