### 2026-03-18 95fe199
Run notebook tests only on pull-requests and schedule. Updated version-numbering.
