### 2026-03-18 f6f79c8
Improved completeness of integration tests and warning-handling (#25)
* Fixed bug encountered with numpy2.4 and typo for "delta_wave". Updated support to python>=3.10. Added weekly tests.

* Fixed import issues in getting_started and pressure_broadening notebooks

* Updated HITRAN cia urls and fixed relative directories in cia notebook

* Added note column-widths in Borysow cia files

* Custom pyROXWarning, to be ignored during tests

* Tests of tutorial notebooks

* More complete end-to-end integration tests

* Run notebook tests only on pull-requests and schedule. Updated version-numbering.
