+++ pyROX/__init__.py
@@ -1,5 +1,5 @@
 # Initialize the package
-__version__ = '1.0.1'
+__version__ = '1.0.2'
 
 from . import utils
 from .utils import sc
