### 2026-03-18 samderegt/pRT_opacities#23 @samderegt
Thanks for this suggestion, I have implemented it now.
