### 2026-03-24 1f019ea
Simplified parallelisation through joblib (#27)
* Simplified parallelisation through joblib

* Required package joblib
