+++ pyROX/cross_sections.py
@@ -417,6 +417,11 @@ class CrossSections:
         self.database = getattr(self.config, 'database', None)
         self.species  = getattr(self.config, 'species', None)
 
+        self.N_CPUs = getattr(self.config, 'N_CPUs', None)
+        if self.N_CPUs is None:
+            import os
+            self.N_CPUs = os.cpu_count()
+
         # Wavenumber/wavelength grid
         self._setup_nu_grid()
         
