### 2026-03-24 5737649
Maintain progress bar information with parallelisation
