### 2026-03-24 f7ed081
Fixed integer display of N_lines_computed
