### 2026-03-24 samderegt/pRT_opacities#21 @samderegt
Hi @kyleaoman, I have now added running the notebooks as part of the scheduled and pull-request tests (they aren't run at every push since they can take a few minutes to complete). In this case, it was indeed as easy as running `pytest --nbmake **/*ipynb`.

As mentioned in the new issue #28, I also have plans for implementing unit tests for some critical parts of the code.
