### 2026-03-26 #10081 @samaloney
@arjunsavel after speaking to other editors and the EiC we can waive the COI in this specific case so please do complete your review, thanks again.
