### 2026-03-31 samderegt/pRT_opacities#29 @arjunsavel
JOSS paper comments
This issue is related to [my JOSS review](https://github.com/openjournals/joss-reviews/issues/10081).

The manuscript for this project is in excellent condition. My minor notes are below:

- In the Summary section, I would briefly define “opacities” — this will make the abstract-equivalent accessible to a wider audience.
- Similarly, defining “spectral line lists” could help.
- "Opacities" and "cross-sections" are used interchangeably — this is OK, but perhaps define each of these on first usage.
- Perhaps cite one of the Niraula papers for line list inaccuracies for space-based spectroscopy.
- I would also cite Simmonin+ 2025, which demonstrated the sensitivity of a VO detection to the line list used in high-resolution spectroscopy.
- Software design: why is this package written in python? Given that python is slower than, e.g., C, brief justification (such as accessibility) would be helpful. (Though you may already be familiar, I'll note that JAX and numba are popular tools used to accelerate Python code!)
- please provide a brief description re: how this code differs from Cthulhu, in e.g., design, purpose.
- It would be useful to describe typical runtime for code evaluation in the manuscript.
- Please check that citation order is chronological throughout.
- Please briefly describe the different pressure broadening prescriptions.

Typos:
- existing open source --> existing open-source
- which cannot be achieved with equal --> , which cannot be achieved with equal
- and this reach is currently expanding to other research groups as well -->, and this reach is currently expanding to other research groups, as well
