### 2026-03-31 #10081 @arjunsavel
@samaloney I've completed my initial review! My assessment is that the authors have provided a useful, well-documented code. The code follows recent best practices outlined in the literature, and the corresponding code paper is nice and pithy. The code itself is well-commented, and I don't see any obvious pathological patterns in design or implementation. I'm inclined to begin using this code in my own research.

My comments on the paper have been left as an issue in the code repository: https://github.com/samderegt/pyROX/issues/29

For the authors (@samderegt), I do have a few questions. I don't think they're issues, per se, though answers might go in the documentation.
- is there a straightforward way to apply highly customized broadening? I'm thinking, for instance, of the Allard+2019 alkali wings.
- a brief comment for future directions of this software might be useful in the paper, documentation, or readme.
